+++ vllm/model_executor/layers/pool/heads/base.py
@@ -6,11 +6,7 @@ from typing import TypeAlias
 import torch
 import torch.nn as nn
 
-from vllm.model_executor.layers.pool.methods import (
-    TokenPoolingMethodOutput,
-    TokenwisePoolingMethodOutputItem,
-)
-from vllm.pooling_params import PoolingParams
+from vllm.model_executor.layers.pool.methods import PoolingMethodOutput
 from vllm.v1.pool.metadata import PoolingMetadata
 
 TokenPoolerHeadOutput: TypeAlias = torch.Tensor | list[torch.Tensor]
@@ -19,25 +15,14 @@ TokenPoolerHeadOutput: TypeAlias = torch.Tensor | list[torch.Tensor]
 TokenwisePoolerHeadOutput: TypeAlias = torch.Tensor | None
 """Applicable to pooling strategies that output multiple tokens."""
 
+PoolerHeadOutput: TypeAlias = TokenPoolerHeadOutput | TokenwisePoolerHeadOutput
 
-class TokenPoolerHead(nn.Module, ABC):
+
+class PoolerHead(nn.Module, ABC):
     @abstractmethod
     def forward(
         self,
-        pooled_data: TokenPoolingMethodOutput,
+        pooled_data: PoolingMethodOutput,
         pooling_metadata: PoolingMetadata,
-    ) -> TokenPoolerHeadOutput:
+    ) -> PoolerHeadOutput:
         raise NotImplementedError
-
-
-class TokenwisePoolerHead(nn.Module, ABC):
-    @abstractmethod
-    def forward(
-        self,
-        pooled_data: TokenwisePoolingMethodOutputItem,
-        pooling_param: PoolingParams,
-    ) -> TokenwisePoolerHeadOutput:
-        raise NotImplementedError
-
-
-PoolerHead: TypeAlias = TokenPoolerHead | TokenwisePoolerHead
